% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianNormal.R
\name{ad.propSd_random}
\alias{ad.propSd_random}
\title{Adaptation For The Proposal Variance}
\usage{
ad.propSd_random(chain, propSd, iteration, lower = 0.3, upper = 0.6,
  delta.n = function(n) min(0.1, 1/sqrt(n)))
}
\arguments{
\item{chain}{matrix of Markov chain samples}

\item{propSd}{old proposal standard deviation}

\item{iteration}{number of current iteration}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{delta.n}{function for adding/subtracting from the log propSd}
}
\description{
Calculation of new proposal standard deviation for the random effects
}
\references{
Rosenthal, J. S. (2011). Optimal proposal distributions and adaptive MCMC. Handbook of Markov Chain Monte Carlo, 93-112.
}
