% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Bayes.fit-class}
\alias{Bayes.fit-class}
\title{S4 class for the Bayesian estimation results}
\description{
S4 class for the Bayesian estimation results
}
\section{Slots}{

\describe{
\item{\code{sigma2}}{vector of posterior samples for \eqn{\sigma^2}}

\item{\code{mu}}{matrix of posterior samples for \eqn{\mu}}

\item{\code{omega}}{matrix of posterior samples for \eqn{\omega}}

\item{\code{alpha}}{matrix of posterior samples for \eqn{\alpha}}

\item{\code{beta}}{matrix of posterior samples for \eqn{\beta}}

\item{\code{random}}{1, 2 or c(1,2)}

\item{\code{burnIn}}{proposal for the burn-in phase}

\item{\code{thinning}}{proposal for the thinning rate}

\item{\code{model}}{'OU' or 'CIR'}

\item{\code{prior}}{list of prior values, input variable or calculated by the first 10\% of series}

\item{\code{times}}{vector of observation times, storage of input variable}

\item{\code{X}}{matrix of observations, storage of input variable}

\item{\code{ind.4.prior}}{indices of series used for the prior parameter calculation, if prior knowledge is availabe it is set to M+1}
}}

