% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dcCIR2}
\alias{dcCIR2}
\title{Likelihood Function For The CIR Model}
\usage{
dcCIR2(x, t, x0, theta, log = FALSE)
}
\arguments{
\item{x}{current observation}

\item{t}{time of observation}

\item{x0}{starting point, i.e. observation in time 0}

\item{theta}{parameter \eqn{(\alpha, \beta, \sigma)}}

\item{log}{logical(1) if TRUE, log likelihood}
}
\description{
Likelihood
}
\references{
Iacus, S. M. (2008). Simulation and Inference for Stochastic Differential Equations.
}

