% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BayesianNormal.R
\name{diagnostic}
\alias{diagnostic}
\title{Calcucation of burn-in phase and thin rate}
\usage{
diagnostic(results, random)
}
\arguments{
\item{results}{Bayes.fit class object}

\item{random}{one out of 1, 2, c(1,2)}
}
\description{
Proposal for burn-in and thin rate
}

