% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eigenvaluesV.R
\name{eigenvaluesV}
\alias{eigenvaluesV}
\title{Matrix Of Eigenvalues Of A List Of Symetric Matrices}
\usage{
eigenvaluesV(V)
}
\arguments{
\item{V}{list of matrices Vj}
}
\value{
\item{eigenvalues}{Matrix of 2 rows and as much columns as matrices V}
}
\description{
Computation of the eigenvalues of each matrix Vj in the case of two random effects (random =c(1,2)), done via \code{eigen}
}
\references{
See Bidimensional random effect estimation in mixed stochastic differential model, C. Dion and V. Genon-Catalot,  \emph{Stochastic Inference for Stochastic Processes 2015, Springer Netherlands}, \bold{1--28}
}

