% Generated by roxygen2 (4.0.1): do not edit by hand
\name{EstParamNormal}
\alias{EstParamNormal}
\title{Maximization Of The Log Likelihood In Mixed Stochastic Differential Equations}
\usage{
EstParamNormal(U, V, K, random, estim.fix, fixed = 0)
}
\arguments{
\item{U}{matrix of M sufficient statistics U}

\item{V}{list of the M sufficient statistics matrix V}

\item{K}{number of times of observations}

\item{random}{random effects in the drift: 1 if one additive random effect, 2 if one multiplicative random effect or c(1,2) if 2 random effects.}

\item{estim.fix}{1 if the fixed parameter is estimated, when random 1 or 2 , 0 otherwise}

\item{fixed}{value of the fixed parameter if known (not estimated)}
}
\value{
\item{mu}{estimated value of the mean}
\item{Omega}{estimated value of the variance}
\item{BIChere}{BIC indicator}
\item{AIChere}{AIC indicator}
}
\description{
Maximization of the loglikelihood of the mixed SDE with Normal distribution of the random effects
 \eqn{dXj(t)= (\alpha_j- \beta_j Xj(t))dt + \sigma a(Xj(t)) dWj(t)}, done with \code{\link[=mixedsde]{likelihoodNormal}}
}

