% Generated by roxygen2 (4.0.1): do not edit by hand
\name{likelihoodNormal}
\alias{likelihoodNormal}
\title{Computation Of The Log Likelihood In Mixed Stochastic Differential Equations}
\usage{
likelihoodNormal(mu, omega, U, V, estimphi, random)
}
\arguments{
\item{mu}{current value of the mean of the normal distribution}

\item{omega}{current value of the standard deviation of the normal distribution}

\item{U}{vector of the M sufficient statistics U (see \code{\link{UV}})}

\item{V}{vector of the M sufficient statistics V (see \code{\link{UV}})}

\item{estimphi}{vector or matrix of estimators of the random effects}

\item{random}{random effects in the drift: 1 if one additive random effect, 2 if one multiplicative random effect or c(1,2) if 2 random effects.}
}
\value{
\item{L}{value of -2 x loglikelihood}
}
\description{
Computation of -2 loglikelihood of the mixed SDE with Normal distribution of the random effects
 \eqn{dXj(t)= (\alpha_j- \beta_j Xj(t))dt + \sigma a(Xj(t)) dWj(t)}.
}
\references{
Maximum likelihood estimation for stochastic differential equations with random effects, M. Delattre, V. Genon-Catalot and A. Samson, \emph{Scandinavian Journal of Statistics 2012}, Vol 40, \bold{322--343}
}

