% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixedsde.fit.R
\docType{methods}
\name{out,Bayes.fit-method}
\alias{out,Bayes.fit-method}
\title{Transfers the class object Bayes.fit to a list}
\usage{
\S4method{out}{Bayes.fit}(x)
}
\arguments{
\item{x}{Bayes.fit class}
}
\description{
Method for the S4 class Bayes.fit
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: an R package to fit mixed stochastic differential equations.
}

