% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{out,Bayes.pred-method}
\alias{out,Bayes.pred-method}
\title{Transfers the class object Bayes.pred to a list}
\usage{
\S4method{out}{Bayes.pred}(x)
}
\arguments{
\item{x}{Bayes.pred class}
}
\description{
Method for the S4 class Bayes.pred
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: an R package to fit mixed stochastic differential equations.
}

