% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{out,Freq.fit-method}
\alias{out,Freq.fit-method}
\title{Transfers the class object Freq.fit to a list}
\usage{
\S4method{out}{Freq.fit}(x)
}
\arguments{
\item{x}{Freq.fit class}
}
\description{
Method for the S4 class Freq.fit
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: an R package to fit mixed stochastic differential equations.
}

