% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedsde.fit.R
\docType{methods}
\name{plot,Bayes.pred,ANY-method}
\alias{plot,Bayes.pred,ANY-method}
\title{Plot method for the Bayesian prediction class object}
\usage{
\S4method{plot}{Bayes.pred,ANY}(x, newwindow = FALSE, plot.legend = TRUE,
  ylim, xlab = "times", ylab = "X", col = 3, lwd = 2, ...)
}
\arguments{
\item{x}{Bayes.fit class}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}

\item{plot.legend}{logical(1)}

\item{ylim}{optional}

\item{xlab}{optional, default 'times'}

\item{ylab}{optional, default 'X'}

\item{col}{color for the prediction intervals, default 3}

\item{lwd}{linewidth for the prediction intervals, default 2}

\item{...}{optional plot parameters}
}
\description{
Plot method for the S4 class Bayes.pred
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: a R package to fit mixed stochastic differential equations.
}
