% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedsde.fit.R
\docType{methods}
\name{plot2compare,Bayes.fit-method}
\alias{plot2compare,Bayes.fit-method}
\title{Comparing plot method plot2compare for three Bayesian estimation class objects}
\usage{
\S4method{plot2compare}{Bayes.fit}(x, y, z, names, true.values,
  reduced = TRUE, newwindow = FALSE)
}
\arguments{
\item{x}{Bayes.fit class}

\item{y}{Bayes.fit class}

\item{z}{Bayes.fit class (optional)}

\item{names}{character vector of names for x, y and z}

\item{true.values}{list of parameters to compare with the estimations, if available}

\item{reduced}{logical(1), if TRUE, the chains are reduced with the burn-in and thin rate}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}
}
\description{
Comparison of the posterior densities for up to three S4 class Bayes.fit objects
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: a R package to fit mixed stochastic differential equations.
}
