% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mixedsde.fit.R
\docType{methods}
\name{plot2compare,Bayes.pred-method}
\alias{plot2compare,Bayes.pred-method}
\title{Comparing plot method plot2compare for three Bayesian prediction class objects}
\usage{
\S4method{plot2compare}{Bayes.pred}(x, y, z, newwindow = FALSE,
  plot.legend = TRUE, names, ylim, xlab = "times", ylab = "X", ...)
}
\arguments{
\item{x}{Bayes.pred class}

\item{y}{Bayes.pred class}

\item{z}{Bayes.pred class (optional)}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}

\item{plot.legend}{logical(1), if TRUE, a legend is added}

\item{names}{character vector with names for the three objects appearing in the legend}

\item{ylim}{optional}

\item{xlab}{optional, default 'times'}

\item{ylab}{optional, default 'X'}

\item{...}{optional plot parameters}
}
\description{
Comparison of the results for up to three S4 class Bayes.pred objects
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: an R package to fit mixed stochastic differential equations.
}

