% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedsde.fit.R
\name{plot2compare}
\alias{plot2compare}
\title{Comparing plot method}
\usage{
plot2compare(x, y, z, ...)
}
\arguments{
\item{x}{Bayes.fit or Bayes.pred class}

\item{y}{Bayes.fit or Bayes.pred class}

\item{z}{Bayes.fit or Bayes.pred class (optional)}

\item{...}{other parameters}
}
\description{
Method for classes
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: a R package to fit mixed stochastic differential equations.
}
