% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pred}
\alias{pred}
\title{Prediction method}
\usage{
pred(x, Xtrue, estim.method, times, level = 0.05, newwindow = FALSE,
  plot.pred = TRUE, ...)
}
\arguments{
\item{x}{Freq.fit or Bayes.fit class}

\item{Xtrue}{observed data}

\item{estim.method}{nonparam or paramML}

\item{times}{observation times}

\item{level}{alpha for the predicion intervals, default 0.05}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}

\item{plot.pred}{logical(1), if TRUE, the results are depicted grafically}

\item{...}{optional plot parameters}
}
\description{
Prediction
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: an R package to fit mixed stochastic differential equations.
}

