% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{print,Bayes.fit-method}
\alias{print,Bayes.fit-method}
\title{Print of acceptance rates of the MH steps}
\usage{
\S4method{print}{Bayes.fit}(x)
}
\arguments{
\item{x}{Bayes.fit class}
}
\description{
Method for the S4 class Bayes.fit
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: an R package to fit mixed stochastic differential equations.
}

