% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{print,Freq.fit-method}
\alias{print,Freq.fit-method}
\title{Description of print}
\usage{
\S4method{print}{Freq.fit}(x)
}
\arguments{
\item{x}{Freq.fit class}
}
\description{
Method for the S4 class Freq.fit
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: an R package to fit mixed stochastic differential equations.
}

