% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{valid,Bayes.fit-method}
\alias{valid,Bayes.fit-method}
\title{Validation of the chosen model.}
\usage{
\S4method{valid}{Bayes.fit}(x, Mrep = 100, newwindow = FALSE,
  plot.valid = TRUE, numj, ...)
}
\arguments{
\item{x}{Bayes.fit class}

\item{Mrep}{number of trajectories to be drawn}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}

\item{plot.valid}{logical(1), if TRUE, the results are depicted grafically}

\item{numj}{optional number of series to be validated}

\item{...}{optional plot parameters}
}
\description{
Validation of the chosen model. For the index numj, Mrep=100 new trajectories are simulated
with the value of the estimated random effect number numj. Two plots are given: on the left the simulated trajectories and the true one (red)
and one the left the corresponding qq-plot for each time.
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: an R package to fit mixed stochastic differential equations.
}

