% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixedsde.fit.R
\docType{methods}
\name{valid,Freq.fit-method}
\alias{valid,Freq.fit-method}
\title{Validation of the chosen model.}
\usage{
\S4method{valid}{Freq.fit}(x, Xtrue, times, Mrep = 100, newwindow = FALSE,
  plot.valid = TRUE, numj, ...)
}
\arguments{
\item{x}{Freq.fit class}

\item{Xtrue}{observed data}

\item{times}{observation times}

\item{Mrep}{number of trajectories to be drawn}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}

\item{plot.valid}{to be added}

\item{numj}{to be added}

\item{...}{optional plot parameters}
}
\description{
Validation of the chosen model. For the index numj, Mrep=100 new trajectories are simulated
with the value of the estimated random effect number numj. Two plots are given: on the left the simulated trajectories and the true one (red)
and one the left the corresponding qq-plot for each time.
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: an R package to fit mixed stochastic differential equations.
}

