% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedsde.fit.R
\name{valid}
\alias{valid}
\title{Validation of the chosen model.}
\usage{
valid(x, ...)
}
\arguments{
\item{x}{Freq.fit or Bayes.fit class}

\item{...}{other optional parameters}
}
\description{
Validation of the chosen model. For the index numj, Mrep=100 new trajectories are simulated
with the value of the estimated random effect number numj. Two plots are given: on the left the simulated trajectories and the true one (red)
and one the left the corresponding qq-plot for each time.
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: a R package to fit mixed stochastic differential equations.
}
