% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedsde.fit.R
\docType{class}
\name{Bayes.pred-class}
\alias{Bayes.pred-class}
\title{S4 class for the Bayesian prediction results}
\description{
S4 class for the Bayesian prediction results
}
\section{Slots}{

\describe{
\item{\code{phi.pred}}{matrix of predictive samples for the random effect}

\item{\code{Xpred}}{matrix of predictive samples for observations}

\item{\code{coverage.rate}}{amount of covering prediction intervals}

\item{\code{qu.u}}{upper prediction interval bound}

\item{\code{qu.l}}{lower prediction interval bound}

\item{\code{estim}}{list of Bayes.fit object entries, storage of input variable}
}}

