% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BayesianNormal.R
\name{diagnostic.own}
\alias{diagnostic.own}
\title{Calcucation of burn-in phase and thin rate}
\usage{
diagnostic.own(chain, dependence = 0.8, m = 10)
}
\arguments{
\item{chain}{vector of Markov chain samples}

\item{dependence}{allowed dependence for the chain}

\item{m}{number of blocks}
}
\description{
Proposal for burn-in and thin rate
}

