% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diagnostic}
\alias{diagnostic}
\title{Calcucation Of Burn-in Phase And Thinning Rate}
\usage{
diagnostic(results, random)
}
\arguments{
\item{results}{Bayes.fit class object}

\item{random}{one out of 1, 2, c(1,2)}
}
\description{
Proposal for burn-in and thin rate
}

