% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mixedsde.fit.R
\docType{class}
\name{Freq.fit-class}
\alias{Freq.fit-class}
\title{S4 class for the frequentist estimation results}
\description{
S4 class for the frequentist estimation results
}
\section{Slots}{

\describe{
\item{\code{model}}{character 'OU' or 'CIR'}

\item{\code{random}}{numeric 1, 2, or c(1,2)}

\item{\code{fixed}}{numeric value of the fixed effect if there is one}

\item{\code{gridf}}{matrix of values on which the estimated is done}

\item{\code{mu}}{numeric MLE estimator for parametric approach}

\item{\code{omega}}{numeric  MLE estimator for parametric approach}

\item{\code{cutoff}}{value of the cutoff if there is one}

\item{\code{sigma2}}{numeric estimated value of \eqn{\sigma^2}}

\item{\code{estimf.trunc}}{matrix estimator of the density of \eqn{\phi} for the truncated estimateur of the random effects}

\item{\code{estimphi.trunc}}{matrix truncated estimator of the random effects}

\item{\code{index}}{index of the used trajectories}

\item{\code{estimphi}}{matrix of the estimator of the random effects}

\item{\code{estimf}}{estimator of the density of \eqn{\phi}}

\item{\code{estim.fixed}}{estimator of the fixed parameter if option estim.fix = 1}

\item{\code{estim.fix}}{1 if the user asked for the estimation of fixed parameter}

\item{\code{bic}}{numeric bic}

\item{\code{aic}}{numeric aic}

\item{\code{times}}{vector of observation times, storage of input variable}

\item{\code{X}}{matrix of observations, storage of input variable}
}}

