% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedsde.fit.R
\name{out}
\alias{out}
\title{Transfers the class object to a list}
\usage{
out(x)
}
\arguments{
\item{x}{Freq.fit, Bayes.fit or Bayes.pred class}
}
\description{
Method for the S4 classes
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: a R package to fit mixed stochastic differential equations.
}
