% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mixedsde.fit.R
\docType{methods}
\name{plot,Freq.fit,ANY-method}
\alias{plot,Freq.fit,ANY-method}
\title{Plot method for the frequentist estimation class object}
\usage{
\S4method{plot}{Freq.fit,ANY}(x, newwindow = FALSE, ...)
}
\arguments{
\item{x}{Freq.fit class}

\item{newwindow}{logical(1), if TRUE, a new window is opened for the plot}

\item{...}{optional plot parameters}
}
\description{
Plot method for the S4 class Freq.fit
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: a R package to fit mixed stochastic differential equations.
}

