% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.compare}
\alias{plot.compare}
\title{Comparing plot method}
\usage{
plot.compare(x, y, z, ...)
}
\arguments{
\item{x}{Bayes.fit or Bayes.pred class}

\item{y}{Bayes.fit or Bayes.pred class}

\item{z}{Bayes.fit or Bayes.pred class}

\item{...}{other parameters}
}
\description{
Method for classes
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: an R package to fit mixed stochastic differential equations.
}

