% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pred}
\alias{pred}
\title{Prediction method}
\usage{
pred(x, ...)
}
\arguments{
\item{x}{Freq.fit or Bayes.fit class}

\item{...}{other optional parameters}
}
\description{
Prediction
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: an R package to fit mixed stochastic differential equations.
}

