% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mixedsde.fit.R
\docType{methods}
\name{summary,Bayes.fit-method}
\alias{summary,Bayes.fit-method}
\title{Short summary of the results of class object Bayes.fit}
\usage{
\S4method{summary}{Bayes.fit}(object, level = 0.05, burnIn, thinning)
}
\arguments{
\item{object}{Bayes.fit class}

\item{level}{default is 0.05}

\item{burnIn}{optional}

\item{thinning}{optional}
}
\description{
Method for the S4 class Bayes.fit
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: a R package to fit mixed stochastic differential equations.
}

