% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{summary,Freq.fit-method}
\alias{summary,Freq.fit-method}
\title{Short summary of the results of class object Freq.fit}
\usage{
\S4method{summary}{Freq.fit}(object)
}
\arguments{
\item{object}{Freq.fit class}
}
\description{
Method for the S4 class Freq.fit
}
\references{
Dion, C., Hermann, S. and Samson, A. (2016). Mixedsde: an R package to fit mixed stochastic differential equations.
}

